backend/requirements.txt
django==4.2.*
djangorestframework==3.14.*
django-cors-headers==4.3.*
reportlab==4.1.*
gunicorn==21.2.*
psycopg2-binary==2.9.*
dj-database-url==2.1.*
whitenoise==6.6.*
python-decouple==3.8
